requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreschema==0.0.4
croniter==0.3.31
Django==2.2.11
django-filter==2.2.0
django-redis==3.5.2
django-redis-cache==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
flower==0.9.4
humanize==0.5.1
idna==2.9
importlib-metadata==1.6.0
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.8
Markdown==3.2.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
tornado==5.1.1
uritemplate==3.0.1
urllib3==1.25.8
uuid==1.30
vine==1.3.0
zipp==1.2.0
